Use packages
There is no particular need to rely on filename convention (i.e.
`template.cue` and `instances.cue`). In fact, such model breaks
compatibility with other usage of CUE, and generally packages are
more convenient to use. With this change a `kg` template can be
evaluated by CUE CLI, and it can use other packages that may be
of use outside of `kg`.